Extract stream.hpp from blocks.hpp
